+++ packages/veui/demo/cases/Select.vue
@@ -3,7 +3,9 @@
     <h1><code>&lt;veui-select&gt;</code></h1>
     <section>
       <h2>默认样式：</h2>
-      <veui-select v-bind="attrs" v-model="defaultValue1" placeholder="全部" clearable></veui-select>
+      <veui-select v-bind="attrs" v-model="defaultValue1" placeholder="全部" clearable :filter="filter">
+        <div slot="before"><veui-input class="filter" ui="tiny" v-model="keyword"/></div>
+      </veui-select>
     </section>
     <section>
       <h2>显示已选图标样式：</h2>
@@ -26,7 +28,7 @@
       <veui-select v-bind="attrs" v-model="defaultValue4">
         <template slot="option" slot-scope="props">
           <span class="veui-option-label-text">{{ props.label }}</span>
-          <icon name="eye"></icon>
+          <veui-icon name="eye"/>
         </template>
       </veui-select>
     </section>
@@ -34,8 +36,35 @@
       <h2>Slot 样式 3：</h2>
       <veui-select v-bind="attrs" v-model="defaultValue1">
         <template slot="option" slot-scope="props">
-          <radio :checked="props.selected">{{ props.label }}</radio>
+          <veui-radio :checked="props.selected">{{ props.label }}</veui-radio>
         </template>
+      </veui-select>
+    </section>
+    <section>
+      <h2>多级样式：</h2>
+      <veui-select v-model="hero">
+        <veui-option-group label="中国" position="popout">
+          <veui-option-group label="互联网" position="popout">
+            <veui-option value="baidu" label="百度"/>
+            <veui-option value="alibaba" label="阿里巴巴"/>
+            <veui-option value="tencent" label="腾讯"/>
+          </veui-option-group>
+          <veui-option-group label="通信" position="popout">
+            <veui-option value="telecom" label="中国电信"/>
+            <veui-option value="mobile" label="中国移动"/>
+          </veui-option-group>
+        </veui-option-group>
+        <veui-option-group label="美国" position="popout">
+          <veui-option-group label="互联网" position="popout">
+            <veui-option value="google" label="Google"/>
+            <veui-option value="facebook" label="Facebook"/>
+            <veui-option value="twitter" label="Twitter"/>
+          </veui-option-group>
+          <veui-option-group label="通信" position="popout">
+            <veui-option value="atnt" label="AT&T"/>
+            <veui-option value="verizon" label="Verizon"/>
+          </veui-option-group>
+        </veui-option-group>
       </veui-select>
     </section>
     <section style="margin-top:500px;">
@@ -61,7 +90,7 @@
         }">
         <template slot="option" slot-scope="props">
           <span class="veui-option-label-text veui-option-custom-label">{{ props.label }}</span>
-          <icon name="gift"/>
+          <veui-icon name="gift"/>
         </template>
       </veui-select>
     </section>
@@ -71,7 +100,7 @@
     </section>
     <section>
       <h2>内联组件嵌套：</h2>
-      <veui-select v-model="defaultValue10">
+      <veui-select v-model="defaultValue10" clearable>
         <veui-option-group label="字母">
           <veui-option value="A" label="选项 A"/>
           <veui-option value="B" label="选项 B"/>
@@ -89,7 +118,7 @@
 
 <script>
 import bus from '../bus'
-import { Icon, Select, Option, OptionGroup, Radio } from 'veui'
+import { Icon, Select, Option, OptionGroup, Radio, Input } from 'veui'
 import type from 'veui/managers/type'
 import 'vue-awesome/icons/eye'
 import 'vue-awesome/icons/gift'
@@ -100,11 +129,17 @@ export default {
     'veui-select': Select,
     'veui-option': Option,
     'veui-option-group': OptionGroup,
-    'radio': Radio,
-    'icon': Icon
+    'veui-radio': Radio,
+    'veui-icon': Icon,
+    'veui-input': Input
   },
   data () {
     return {
+      filter: ({ label }) => {
+        return label.indexOf(this.keyword) !== -1
+      },
+      keyword: '',
+      hero: null,
       disabled: true,
       selected: true,
       icon: true,
@@ -122,22 +157,22 @@ export default {
         name: 'age',
         readonly: false,
         disabled: false,
-        selecticon: false,
+        clearable: true,
         options: [
           {
-            label: '下拉选项1',
+            label: '下拉选项111',
             value: 1
           },
           {
-            label: '下拉选项下拉选项下拉选项22',
+            label: '下拉选项下拉选项下拉选项112',
             value: 2
           },
           {
-            label: '下拉选项3',
+            label: '下拉选项122',
             value: 3
           },
           {
-            label: '下拉选项4',
+            label: '下拉选项222',
             value: 4
           }
         ]
@@ -148,9 +183,11 @@ export default {
         readonly: false,
         disabled: false,
         selecticon: false,
+        clearable: true,
         options: [
           {
             label: '下拉选项1',
+            position: 'popout',
             options: [
               {
                 label: '子选项1',
@@ -258,6 +295,7 @@ export default {
     margin-left: 5px;
   }
 }
+
 .veui-option-custom {
   position: relative;
   padding-left: 10px;
@@ -272,5 +310,10 @@ export default {
   &-label {
     color: rgb(80, 170, 39);
   }
+}
+
+.filter {
+  width: ~"calc(100% - 10px)";
+  margin: 5px;
 }
 </style>