+++ packages/veui/src/components/Label.vue
@@ -1,13 +1,15 @@
 <template>
-<label class="veui-label" @click="findLabeledInput"><slot></slot></label>
+<label class="veui-label" :ui="ui" @click="findLabeledInput"><slot/></label>
 </template>
 
 <script>
-import { isFunction, get, includes } from 'lodash'
-import { getTypedAncestor } from '../utils/helper'
+import { isFunction, get } from 'lodash'
+import { getTypedAncestor, isType } from '../utils/helper'
+import ui from '../mixins/ui'
+
 export default {
   name: 'veui-label',
-
+  mixins: [ui],
   methods: {
     findLabeledInput () {
       if (window.getSelection().toString().length) {
@@ -16,7 +18,7 @@ export default {
       let ancestor = getTypedAncestor(this, 'field')
       if (ancestor) {
         let target = ancestor.$children.filter(child => child !== this)[0]
-        while (target && !includes(get(target, '$options.uiTypes', []), 'input')) {
+        while (target && !isType(target, 'input')) {
           target = get(target, '$children[0]')
         }