+++ packages/veui/src/components/Select/OptionGroup.vue
@@ -1,70 +1,55 @@
-<template>
-<div :class="{
-  'veui-option-group': true,
-  'veui-option-group-unlabelled': !label
-}">
-  <div v-if="label" class="veui-option-group-label">
-    <slot name="label" :label="label">{{ label }}</slot>
-  </div>
-  <template v-if="options">
-    <template
-      v-for="(option, i) in options">
-      <veui-option
-        v-if="!option.options"
-        v-bind="option"
-        :ui="ui"
-        :key="option.value">
-        <slot v-if="$scopedSlots.option" name="option" v-bind="option" :selected="option.value === value"></slot>
-        <template v-if="$scopedSlots['option-label']" slot="label">
-          <slot name="option-label" v-bind="option" :selected="option.value === value">{{ option.label }}</slot>
-        </template>
-      </veui-option>
-      <veui-option-group
-        v-else
-        v-bind="option"
-        :ui="ui"
-        :key="i">
-        <template v-if="$scopedSlots.label" slot="label" slot-scope="group">
-          <slot name="label" v-bind="group">{{ group.label }}</slot>
-        </template>
-        <template v-if="$scopedSlots.option" slot="option" slot-scope="option">
-          <slot name="option" v-bind="option"></slot>
-        </template>
-        <template v-if="$scopedSlots['option-label']" slot="option-label" slot-scope="option">
-          <slot name="option-label" v-bind="option"></slot>
-        </template>
-      </veui-option-group>
-    </template>
-  </template>
-  <template v-else>
-    <slot></slot>
-  </template>
-</div>
-
-</template>
 
 <script>
 import Option from './Option'
+import Overlay from '../Overlay'
+import Icon from '../Icon'
 import ui from '../../mixins/ui'
+import overlay from '../../mixins/overlay'
 import menu from '../../mixins/menu'
 import select from '../../mixins/select'
+import keySelect from '../../mixins/key-select'
+import outside from '../../directives/outside'
+import '../../config/uiTypes'
+import { walk } from '../../utils/data'
+import { pull } from 'lodash'
 
 export default {
   name: 'veui-option-group',
   uiTypes: ['menu'],
-  mixins: [ui, menu, select],
+  mixins: [ui, menu, select, overlay, keySelect],
   components: {
-    'veui-option': Option
+    'veui-option': Option,
+    'veui-overlay': Overlay,
+    'veui-icon': Icon
+  },
+  directives: {
+    outside
   },
   props: {
     label: String,
-    options: {
-      type: Array
+    options: Array,
+    position: {
+      type: String,
+      default: 'inline',
+      validator (val) {
+        return ['inline', 'popout'].indexOf(val) !== -1
+      }
     }
   },
   data () {
     return {
-      items: []
+      items: [],
+      expanded: false,
+      localOverlayOptions: {
+        position: 'right-start',
+        constraints: [
+          {
+            to: 'window',
+            attachment: 'together'
+          }
+        ]
+      },
+      outsideRefs: ['button']
     }
   },
   computed: {
@@ -73,6 +58,153 @@ export default {
     },
     itemIds () {
       return this.items.map(({ id }) => id)
+    },
+    canPopOut () {
+      return !!(this.position === 'popout' && this.items && this.items.length && this.label)
+    }
+  },
+  render () {
+    let content = this.options
+      ? this.options.map((option, i) => {
+        return option.options
+          ? <veui-option-group
+            ui={this.inheritedUi}
+            label={option.label}
+            options={option.options}
+            position={option.position}
+            key={i}
+            scopedSlots={{
+              'group-label': this.$scopedSlots['group-label']
+                ? group => this.$scopedSlots['group-label'](group) || group.label
+                : null,
+              option: this.$scopedSlots.option
+                ? option => this.$scopedSlots.option(option)
+                : null,
+              'option-label': this.$scopedSlots['option-label']
+                ? option => {
+                  return this.$scopedSlots['option-label'](option)
+                }
+                : null
+            }}>
+          </veui-option-group>
+          : <veui-option
+            ui={this.inheritedUi}
+            label={option.label}
+            value={option.value}
+            hidden={option.hidden}
+            key={i}>
+            {
+              this.$scopedSlots.option
+                ? this.$scopedSlots.option(option)
+                : null
+            }
+            <template slot="label">
+              {
+                this.$scopedSlots['option-label']
+                  ? this.$scopedSlots['option-label'](option)
+                  : null
+              }
+            </template>
+          </veui-option>
+      })
+      : this.$slots.default
+
+    let LabelTag = this.canPopOut ? 'button' : 'div'
+
+    return <div
+      class={{
+        'veui-option-group': true,
+        'veui-option-group-unlabelled': !this.label,
+        'veui-option-group-expanded': this.expanded
+      }}
+      ui={this.ui}
+      ref="label">
+      {
+        this.label
+          ? <LabelTag
+            ref="button"
+            class={{
+              'veui-option-group-label': true,
+              'veui-option-group-button': this.canPopOut
+            }}
+            {...this.canPopOut
+              ? {
+                on: {
+                  click: () => {
+                    this.expanded = true
+                  },
+                  keydown: e => {
+                    if (e.key === 'Right' || e.key === 'ArrowRight') {
+                      this.expanded = true
+                      e.stopPropagation()
+                      e.preventDefault()
+                    }
+                  }
+                }
+              }
+              : {}
+            }>
+            <span class="veui-option-label">
+              {
+                this.$scopedSlots.label
+                  ? this.$scopedSlots.label({ label: this.label })
+                  : this.label
+              }
+            </span>
+            {
+              this.canPopOut
+                ? <veui-icon class="veui-option-group-expandable" name={this.icons.expandable}/>
+                : null
+            }
+          </LabelTag>
+          : null
+      }
+      {
+        this.canPopOut
+          ? <veui-overlay
+            ref="overlay"
+            target="button"
+            open={this.expanded}
+            options={this.realOverlayOptions}
+            overlayClass={this.mergeOverlayClass('veui-option-group-box')}
+            autofocus
+            modal>
+            <div
+              ref="box"
+              class="veui-select-options"
+              tabindex="-1"
+              ui={this.ui}
+              {...{
+                directives: [{
+                  name: 'outside',
+                  value: {
+                    refs: this.outsideRefs,
+                    handler: () => {
+                      this.expanded = false
+                    }
+                  }
+                }]
+              }}
+              onKeydown={this.handleKeydown}>
+              {content}
+            </div>
+          </veui-overlay>
+          : content
+      }
+    </div>
+  },
+  watch: {
+    expanded (val) {
+      let box = this.$refs.box
+      let parent = this.menu || this.select
+      while (parent) {
+        if (val) {
+          parent.outsideRefs.push(box)
+        } else {
+          pull(parent.outsideRefs, box)
+        }
+        parent = parent.menu || parent.select
+      }
     }
   },
   methods: {
@@ -90,6 +222,21 @@ export default {
     },
     find (val) {
       return findItemByValue(this.items, val)
+    },
+    relocate () {
+      if (this.canPopOut && this.expanded) {
+        this.$refs.overlay.relocate()
+      }
+    },
+    relocateDeep () {
+      walk(this, child => {
+        if (child.$options.name === this.$options.name) {
+          child.relocate()
+        }
+      }, '$children')
+    },
+    close () {
+      this.expanded = false
     }
   }
 }
@@ -115,4 +262,5 @@ function findItemByValue (items, val) {
   })
   return result
 }
+
 </script>