+++ packages/veui/src/components/Select/Select.vue
@@ -1,56 +1,3 @@
-<template>
-<div class="veui-select" :ui="ui"
-  :class="{
-    'veui-select-empty': value === null,
-    'veui-select-expanded': expanded
-  }">
-  <veui-button
-    class="veui-select-button"
-    :ui="ui"
-    :disabled="realDisabled || realReadonly"
-    @keydown.down.up.prevent="expanded = true"
-    @click="expanded = !expanded"
-    ref="button">
-    <span class="veui-select-label">
-      <slot name="label" :label="label">{{ label }}</slot>
-    </span>
-    <icon class="veui-select-icon" :name="icons[expanded ? 'collapse' : 'expand']"></icon>
-  </veui-button>
-  <veui-overlay
-    v-if="options && expanded || !options"
-    v-show="expanded"
-    target="button"
-    :open="expanded"
-    autofocus
-    modal
-    :options="realOverlayOptions"
-    :overlay-class="overlayClass">
-    <div
-      ref="box"
-      class="veui-select-options"
-      v-outside:button="close"
-      tabindex="-1"
-      :ui="ui"
-      @keydown.esc.stop="expanded = false"
-      @keydown.down.prevent="navigate()"
-      @keydown.up.prevent="navigate(false)">
-      <veui-option-group :options="realOptions" ref="options">
-        <slot></slot>
-        <template v-if="$scopedSlots['group-label']" slot="label" slot-scope="group">
-          <slot name="group-label" v-bind="group"></slot>
-        </template>
-        <template v-if="$scopedSlots.option" slot="option" slot-scope="option">
-          <slot name="option" v-bind="option"></slot>
-        </template>
-        <template v-if="$scopedSlots['option-label']" slot="option-label" slot-scope="option">
-          <slot name="option-label" v-bind="option"></slot>
-        </template>
-      </veui-option-group>
-    </div>
-  </veui-overlay>
-</div>
-</template>
-
 <script>
 import Icon from '../Icon'
 import Button from '../Button'
@@ -58,28 +5,30 @@ import Option from './Option'
 import OptionGroup from './OptionGroup'
 import Overlay from '../Overlay'
 import input from '../../mixins/input'
-import select from '../../mixins/select'
-import icons from '../../mixins/icons'
+import keySelect from '../../mixins/key-select'
+import ui from '../../mixins/ui'
 import overlay from '../../mixins/overlay'
 import dropdown from '../../mixins/dropdown'
 import warn from '../../utils/warn'
+import { walk } from '../../utils/data'
+import { cloneDeep } from 'lodash'
+import '../../config/uiTypes'
 
 export default {
   name: 'veui-select',
   uiTypes: ['select'],
-  mixins: [input, icons, overlay, dropdown, select],
+  mixins: [ui, input, overlay, dropdown, keySelect],
   model: {
     event: 'change'
   },
   components: {
-    Icon,
+    'veui-icon': Icon,
     'veui-button': Button,
     'veui-option': Option,
     'veui-option-group': OptionGroup,
     'veui-overlay': Overlay
   },
   props: {
-    ui: String,
     value: null,
     placeholder: {
       type: String,
@@ -90,23 +39,16 @@ export default {
       default: false
     },
     clearable: Boolean,
-    options: Array
+    options: Array,
+    filter: Function
   },
   data () {
     return {
-      localValue: this.value
+      localValue: this.value,
+      outsideRefs: ['button']
     }
   },
   computed: {
-    realOptions () {
-      if (this.clearable) {
-        return [
-          { label: this.placeholder, value: null },
-          ...this.options
-        ]
-      }
-      return this.options
-    },
     labelMap () {
       return extractOptions(this.options, {})
     },
@@ -114,17 +56,120 @@ export default {
       if (this.value === null) {
         return this.placeholder
       }
-      if (this.labelMap) {
+      if (this.options) {
         return this.labelMap[this.value]
       }
       let option = this.$refs.options.find(this.value)
       return option ? option.label : ''
+    },
+    realOptions () {
+      if (typeof this.filter !== 'function') {
+        return this.options
+      }
+      let filtered = cloneDeep(this.options)
+      walk(filtered, option => {
+        option.hidden = !this.filter(option)
+      })
+      return filtered
     }
+  },
+  render () {
+    return <div
+      class={{
+        'veui-select': true,
+        'veui-select-empty': this.value === null,
+        'veui-select-expanded': this.expanded
+      }}
+      ui={this.ui}>
+      <veui-button
+        ref="button"
+        class="veui-select-button"
+        ui={this.ui}
+        disabled={this.realDisabled || this.realReadonly}
+        onKeydown={this.handleButtonKeydown}
+        onClick={this.handleButtonClick}>
+        <span class="veui-select-label">
+          {
+            this.$scopedSlots.label ? this.$scopedSlots.label({ label: this.label }) : this.label
+          }
+        </span>
+        <veui-icon
+          class="veui-select-icon"
+          name={this.icons[this.expanded ? 'collapse' : 'expand']}/>
+      </veui-button>
+      {
+        this.options && this.expanded || !this.options
+          ? <veui-overlay
+            v-show={this.expanded}
+            target="button"
+            open={this.expanded}
+            autofocus
+            modal
+            options={this.realOverlayOptions}
+            overlay-class={this.overlayClass}
+            onLocate={this.handleRelocate}>
+            <div
+              ref="box"
+              class="veui-select-options"
+              {...{
+                directives: [{
+                  name: 'outside',
+                  value: {
+                    refs: this.outsideRefs,
+                    handler: this.close
+                  }
+                }]
+              }}
+              tabindex="-1"
+              ui={this.ui}
+              onKeydown={this.handleKeydown}>
+              {this.$slots.before}
+              {
+                this.clearable
+                  ? <veui-option value={null} label={this.placeholder}/>
+                  : null
+              }
+              <veui-option-group
+                ref="options"
+                options={this.realOptions}
+                ui={this.ui}
+                scopedSlots={{
+                  label: this.$scopedSlots['group-label']
+                    ? group => this.$scopedSlots['group-label'](group)
+                    : null,
+                  option: this.$scopedSlots.option
+                    ? option => this.$scopedSlots.option(option)
+                    : null,
+                  'option-label': this.$scopedSlots['option-label']
+                    ? option => this.$scopedSlots['option-label'](option)
+                    : null
+                }}>
+                {this.$slots.default}
+              </veui-option-group>
+              {this.$slots.after}
+            </div>
+          </veui-overlay>
+          : null
+      }
+    </div>
   },
   methods: {
     handleSelect (value) {
       this.expanded = false
       this.localValue = value
+    },
+    handleRelocate () {
+      this.$refs.options.relocateDeep()
+    },
+    handleButtonClick () {
+      this.expanded = !this.expanded
+    },
+    handleButtonKeydown (e) {
+      if (e.key === 'Up' || e.key === 'ArrowUp' || e.key === 'Down' || e.key === 'ArrowDown') {
+        this.expanded = true
+        e.stopPropagation()
+        e.preventDefault()
+      }
     }
   },
   watch: {
@@ -140,20 +185,14 @@ export default {
 }
 
 function extractOptions (options, map) {
-  if (!options) {
-    return null
-  }
-  options.forEach(({ label, value, options }) => {
+  walk(options, ({ label, value, options, children }) => {
     if (value != null) {
       if (map[value]) {
-        warn(`Duplicate item value [${value}] for select options.`)
+        warn(`[veui-select] Duplicate item value [${value}] for select options.`)
       }
       map[value] = label
     }
-    if (options) {
-      extractOptions(options, map)
-    }
-  })
+  }, ['options', 'children'])
   return map
 }
 </script>